Mise en forme du code controller
Adaptation du code aux fonctions importées
Création des fichiers sources "utilities"
DEFINE de controller.h importés dans appSACS.h
main.c épuré des variables de test
