+++ STM32/app/inc/controller.h
@@ -9,21 +9,8 @@
 #define APP_INC_CONTROLLER_H_
 
 #include "appSACS.h"
-#include "bsp.h"
-
-/* temporaire, à intégrer dans appSACS.h */
-#define SID1		0x03
-#define LED_ON		0x11
-#define LED_OFF		0x10
-#define LED_TOGGLE	0x12
-
-#define MAX_DATA_SIZE   16
-#define SEND_ERROR	    1
-#define SEND_OK		    0
-#define RECEIVE_ERROR	1
-#define RECEIVE_OK	    0   
-#define ACK             1
-#define NACK            0
+#include "delay.h"
+#include "utilities.h"
 
 /**
  * @brief Run the controller core function
