+++ STM32/app/inc/main.h
@@ -12,7 +12,7 @@
  * printf() and sprintf() from printf-stdarg.c
  */
 
-#define DEBUG 1
+#define DEBUG   1
 
 int my_printf	(const char *format, ...);
 int my_sprintf	(char *out, const char *format, ...);
